Include EITA
Our cooperative is seven years old. We are part of the solidarity economy movement, and we develop systems for movements on agriarian reform, agroecology, popular education, and other. We are seven member working in a true autogestionary model. The mission of our coop is to strengthen the struggles of social movements in the popular field through the development of open source IT and participatory methodologies for their use and appropriation.

Congratulations for this great resource,  and please fell free to share any doubts.